+++ src/algorithm/math/multiple.ts
@@ -5,14 +5,13 @@ export function multiply(num1: string, num2: string): string {
   const w = num2.length;
   const result: number[] = Array(h + w)
     .fill(0);
+  let sum = 0;
 
   for (let i = h - 1; i >= 0; i--) {
     for (let j = w - 1; j >= 0; j--) {
-      const multi = Number(num1[i]) * Number(num2[j]);
-      const currentPosition = i + j + 1;
-      const sum = multi + result[currentPosition];
-      result[currentPosition] = sum % 10;
-      result[currentPosition - 1] += (sum / 10) | 0;
+      sum = Number(num1[i]) * Number(num2[j]) + result[i + j + 1];
+      result[i + j + 1] = sum % 10;
+      result[i + j] += (sum / 10) | 0;
     }
   }