+++ src/algorithm/other/isRectangleCover.ts
@@ -1,6 +1,5 @@
 // https://leetcode-cn.com/problems/perfect-rectangle/
 // 391
-
 export function isRectangleCover(rectangles: number[][]): boolean {
   let leftBottomX = Infinity;
   let leftBottomY = Infinity;